Bug Fixed

+++ database/migrations/2024_05_28_014458_create_reservations_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('reservations', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('users_id');
+            $table->unsignedBigInteger('users_id');
             $table->foreign('users_id')->references('id')->on('users');
             $table->dateTime('reservation_date');
             $table->string('number_of_people');
